+++ src/main/java/atom/marvelores/datagen/ModModelProvider.java
@@ -21,7 +21,6 @@ public class ModModelProvider extends FabricModelProvider {
         blockStateModelGenerator.registerSimpleCubeAll(ModBlocks.VIBRANIUM_ORE);
         blockStateModelGenerator.registerSimpleCubeAll(ModBlocks.VIBRANIUM_DEEPSLATE_ORE);
 
-        blockStateModelGenerator.registerSimpleCubeAll(ModBlocks.CRUSHER);
     }
 
     @Override
@@ -29,6 +28,8 @@ public class ModModelProvider extends FabricModelProvider {
         itemModelGenerator.register(ModItems.VIBRANIUM, Models.GENERATED);
         itemModelGenerator.register(ModItems.RAW_VIBRANIUM, Models.GENERATED);
         itemModelGenerator.register(ModItems.VIBRANIUM_POWDER, Models.GENERATED);
+        itemModelGenerator.register(ModItems.VIBRANIUM_STICK, Models.GENERATED);
+        itemModelGenerator.register(ModItems.VIBRANIUM_CORE, Models.GENERATED);
 
         itemModelGenerator.register(ModItems.VIBRANIUM_SWORD, Models.HANDHELD);
         itemModelGenerator.register(ModItems.VIBRANIUM_PICKAXE, Models.HANDHELD);