+++ src/main/java/atom/marvelores/item/ModItems.java
@@ -14,6 +14,8 @@ public class ModItems {
     public static final Item VIBRANIUM = registerItem("vibranium", new Item(new Item.Settings()));
     public static final Item RAW_VIBRANIUM = registerItem("raw_vibranium", new Item(new Item.Settings()));
     public static final Item VIBRANIUM_POWDER = registerItem("vibranium_powder", new Item(new Item.Settings()));
+    public static final Item VIBRANIUM_STICK = registerItem("vibranium_stick", new Item(new Item.Settings()));
+    public static final Item VIBRANIUM_CORE = registerItem("vibranium_core", new Item(new Item.Settings()));
 
     public static final Item VIBRANIUM_SWORD = registerItem("vibranium_sword",
             new SwordItem(ModToolMaterials.VIBRANIUM,new Item.Settings()